requirements.txt
pyrogram
fastapi
uvicorn
tgcrypto
python-dotenv